Setup to run and profile a test case
The test case is run by sourcing the file scripts/profile.R. This script is very similar to
scripts/runpi.R, except that the input parameters are hard-coded, not passed on the command line,
and the Rprof profiler is used to keep track of where the program spends its time. Profiling reports
are generated by both the summaryRprof() function and the proftools pacakage, which must be
installed. Profiling output is saved alongside the program output, both in the profile/ folder. I
added this folder to .gitignore, so that it is not constantly updated in the repository. Lastly, the
program looks for data in data/SE_Spectra/... that is not available in the repository (a symlink to
an unavailable directory is present instead). I added the subset of this data provided to me by
email from  Alexey Shiklomanov on 1/8/2015 and deleted the symlink.