Added duplicate 'Sample alphaN' code block and tests.
As a precursor to changing the 'Sample alpha*' code blocks, I duplicated the first block and added
code to test that the results from the two duplicate blocks are identical and to time thier
execution. Duplicating the block entailed identifying variables that are changed by the code block
and storing thier initial values, and also setting and resetting the random seed for each duplicate.
The final values after each block are then tested for equivalence. Times are collected using the
proc.time() tool. The duplicate block produces identical results in roughly the same time. The next
step is to make changes, hopefully speeding up the code while retaining identical results.