Test storing likelihood output. Result is ~2x speedup.
In the original version, the likelihood function is called for both guess.error and prev.error. This
is not necessary since the value of prev.error is the same as some past version of guess.error. By
storing the value of likelihood(guess.error, ...), calls to likelihood can be halved. The result is
a ~2x speedup in the speed of the 'Sample alphaN' code block.

I also simplified the guess.alphaN variable (vector reduced to a scalar), but no speedup resulted.

The tests included in the previous commit confirm that the new version gives the exact same results
as the old version.