Prepared to implement 'Sample alphaN' block in C
I removed the original version of the 'Sample alphaN' code block and replaced it with the new
version. There is now an empty slot for a new C version of the code block which I will be writing
and testing against the R version. This commit is to back up against early missteps.