Implemented the 'Sample Alpha N' in C, results match R function exactly.
I implemented a C version of the Sample Alpha N code block that I have been working with in previous
commits. The C version is called using R's .Call functionality and does not have any dependencies.
It is necessary to compile the external routine to a shared library before it can be used. To make
this easy, I have included a simple makefile - just run 'make' to compile and everything will work.
Currently, the code compares results and times for C version and the "new" faster R version of the
code block. As a final step, I will switch this to compare the C version and the original R version.
The "new" faster R version will be accessible through previous commits.