Include comparison of original R, new R, and C code blocks.
I modified inv_bayes.R to run all three versions of the Sample AlphaN code block, and to print the
execution time and equivalence tests for each. The results show that the new R version runs in ~60%
of the time needed by the original R version, and the C version runs in about 35% of the time neded
by the original R version, and all versions produce identical results.

Note that the C implementation of the code block is a bit messy since so many variables have to be
passed in and out, and there is no simple way to this block into a flexible function that could work
for other simlar code blocks without changing the way the data are stored in variables.